# pos-carts/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Carts
  version: 1.0.0
paths:
  /cart:
    get:
      summary: Show item in Cart
      operationId: showCart
      tags:
        - cart
      responses:
        '200':
          description: A paged array of items in cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/{productId}:
    post:
      summary: add an item to Cart
      operationId: addItemToCart
      tags:
        - cart add item
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to retrieve
          schema:
            type: string
      responses:
        '200':
          description: A boolean result of adding operation
          content:
            application/json:
              schema:
                type: boolean
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/total:
    get:
      summary: get total count of cart
      operationId: getTotal
      tags:
        - cart total
      responses:
        '200':
          description: An integer of the cart total count
          content:
            application/json:
              schema:
                type: double
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: double
        image:
          type: string
    Item:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
    Cart:
      type: array
      items:
        $ref: "#/components/schemas/Item"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  